feat: add sign message APDU command handler and UI flow
This commit introduces the sign message functionality with several key improvements:
- Implement handleSign function in sign.c with buffer-based input
- Update dispatcher to support INS_SIGN command
- Add sign validation and UI confirmation flow
- Modify send_response helper to return int for sign response
- Update UI display logic to use a generic validation callback
- Add validate_sign function for message signing and rejection
- Refactor sign handler to use new buffer and error handling approach

+++ src/helper/send_response.h
@@ -6,4 +6,4 @@
 int helper_send_response_public_key();
 int helper_send_response_address();
 uint8_t helper_send_response_sign_transaction();
-uint8_t helper_send_response_sign();
+int helper_send_response_sign();
